# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.CreateItemInput:
    properties:
      description:
        example: This is item1
        type: string
      name:
        description: json:"name"はJSONのキー名を指定し、間違っている場合はエラーを返す
        minLength: 2
        type: string
      price:
        example: 100
        maximum: 999999
        minimum: 1
        type: integer
    required:
    - name
    - price
    type: object
  dto.SignupInput:
    properties:
      email:
        example: john@example.com
        type: string
      name:
        example: John Doe
        type: string
      password:
        example: password123
        minLength: 8
        type: string
    required:
    - email
    - name
    - password
    type: object
  dto.UpdateItemInput:
    properties:
      description:
        type: string
      name:
        minLength: 2
        type: string
      price:
        maximum: 999999
        minimum: 1
        type: integer
      soldOut:
        type: boolean
    type: object
  ent.Item:
    properties:
      description:
        description: Description holds the value of the "description" field.
        type: string
      edges:
        allOf:
        - $ref: '#/definitions/ent.ItemEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the ItemQuery when eager-loading is set.
      id:
        description: ID of the ent.
        type: integer
      name:
        description: Name holds the value of the "name" field.
        type: string
      price:
        description: Price holds the value of the "price" field.
        type: integer
      sold_out:
        description: SoldOut holds the value of the "sold_out" field.
        type: boolean
      user_id:
        description: UserID holds the value of the "user_id" field.
        type: integer
    type: object
  ent.ItemEdges:
    properties:
      user:
        allOf:
        - $ref: '#/definitions/ent.User'
        description: User holds the value of the user edge.
    type: object
  ent.User:
    properties:
      edges:
        allOf:
        - $ref: '#/definitions/ent.UserEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the UserQuery when eager-loading is set.
      id:
        description: ID of the ent.
        type: integer
      login_id:
        description: LoginID holds the value of the "login_id" field.
        type: string
      name:
        description: Name holds the value of the "name" field.
        type: string
    type: object
  ent.UserEdges:
    properties:
      items:
        description: Items holds the value of the items edge.
        items:
          $ref: '#/definitions/ent.Item'
        type: array
    type: object
info:
  contact: {}
  description: This is a sample server for a pet store.
  title: Swagger Example API
  version: "1.0"
paths:
  /api/v1/items/{id}:
    get:
      description: Get item from the database by ID
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get item by ID
      tags:
      - items
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Create a new user with the input payload
      parameters:
      - description: User info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.SignupInput'
      produces:
      - application/json
      responses:
        "201":
          description: created
          schema:
            type: string
        "400":
          description: bad request
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Create a new user
      tags:
      - auth
  /hello:
    get:
      consumes:
      - application/json
      description: do ping
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: liveness probe
      tags:
      - Hello World
  /items:
    get:
      description: Get all items from the database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.Item'
        "500":
          description: Internal Server Error
      summary: Get all items
      tags:
      - items
    post:
      consumes:
      - application/json
      description: Create a new item in the database
      parameters:
      - description: Item information
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/dto.CreateItemInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ent.Item'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Create a new item
      tags:
      - items
  /items/{id}:
    delete:
      description: Delete an item in the database
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Delete an item
      tags:
      - items
    put:
      consumes:
      - application/json
      description: Update an item in the database
      parameters:
      - default: 1
        description: Item ID
        in: path
        name: id
        required: true
        type: integer
      - description: Item information
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateItemInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Update an item
      tags:
      - items
swagger: "2.0"